pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kafka-schema-performance"
version = "0.1.0"
description = "Benchmark harness for Avro, Protobuf, and JSON serialization"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastavro>=1.9.4,<2",
    "protobuf>=5.26,<7",
    "orjson>=3.10,<4",
    "typer>=0.12.3,<1",
    "pyyaml>=6.0.1,<7",
    "zstandard>=0.22,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9",
    "ruff>=0.6.0,<1",
    "mypy>=1.11,<2",
    "black>=24.8,<25",
    "types-PyYAML",
    "grpcio-tools>=1.60,<2",
]

[project.scripts]
ksp-bench = "benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
benchmark = ["fixtures/*"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/benchmark/fixtures/event_pb2.py" = ["ALL"]

[tool.black]
target-version = ["py311"]
line-length = 88
extend-exclude = "/event_pb2\\.py$"

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["benchmark"]

[[tool.mypy.overrides]]
module = "google.protobuf.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fastavro.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "benchmark.fixtures.event_pb2"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
